Change Network

+++ pysrc/common/data_transform_mod.py
@@ -28,13 +28,13 @@ class DataTransform():
         roll_numpy = roll_numpy / np.linalg.norm(roll_numpy)
         #roll_tensor = torch.from_numpy(roll_numpy)
         roll_tensor = torch.tensor(roll_numpy, requires_grad=True)
-        roll_tensor = roll_tensor.long()
+        #roll_tensor = roll_tensor.long()
 
         # pitch: numpy -> tensor
         pitch_numpy = pitch_numpy.astype(np.float32)
         pitch_numpy = pitch_numpy / np.linalg.norm(pitch_numpy)
         #pitch_tensor = torch.from_numpy(pitch_numpy)
         pitch_tensor = torch.tensor(pitch_numpy, requires_grad = True)
-        pitch_tensor = pitch_tensor.long()
+        #pitch_tensor = pitch_tensor.long()
 
         return img_tensor, roll_tensor, pitch_tensor
